pyproject.toml
[tool.poetry]
name = "drfdocker"
version = "0.1.0"
description = ""
authors = ["mist258 <natalia.kolchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.16.1"
psycopg = "^3.3.2"
drf-yasg = "^1.21.10"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
